requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
SQLAlchemy>=2.0
alembic
psycopg[binary]
requests
redis
apscheduler
